refactor: reorgnized file structure
