+++ back-end/src/main/java/edu/brown/cs/student/main/server/Server.java
@@ -1,13 +1,10 @@
-package edu.brown.cs.student.main.server.spotify.data;
+package edu.brown.cs.student.main.server;
 
 import static spark.Spark.after;
 
-import com.beust.ah.A;
-
-import edu.brown.cs.student.main.server.CachedSpotifyData;
-import edu.brown.cs.student.main.server.audioRecognition.audioData.AudioData;
 import edu.brown.cs.student.main.server.handlers.AudioTextHandler;
 import edu.brown.cs.student.main.server.handlers.GetSongHandler;
+import edu.brown.cs.student.main.server.audioRecognition.audioData.AudioData;
 import edu.brown.cs.student.main.server.handlers.AddDislikedSongsHandler;
 import edu.brown.cs.student.main.server.handlers.AddInputSongsHandler;
 import edu.brown.cs.student.main.server.handlers.LyricsHandler;
@@ -15,6 +12,7 @@ import edu.brown.cs.student.main.server.handlers.RecommendationHandler;
 import edu.brown.cs.student.main.server.handlers.ScoreHandler;
 import edu.brown.cs.student.main.server.handlers.TokenHandler;
 import edu.brown.cs.student.main.server.lyrics.data.LyricsData;
+import edu.brown.cs.student.main.server.spotify.data.CachedSpotifyData;
 import edu.brown.cs.student.main.server.spotify.tokens.TokenGenerator;
 import edu.brown.cs.student.main.server.translate.data.AzureTranslateData;
 import edu.brown.cs.student.main.server.translate.data.LibreTranslateData;
